refactor: Add Credit Facility history job

+++ core/credit/src/credit_facility/jobs/history/entry.rs
@@ -48,7 +48,7 @@ pub struct DisbursalExecuted {
 pub struct InterestAccrualsPosted {
     pub cents: UsdCents,
     pub recorded_at: DateTime<Utc>,
-    pub days: i64,
+    pub days: u16,
     pub tx_id: LedgerTxId,
 }
